requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-braces==1.13.0
django-embed-video==1.2.0
djangorestframework==3.9.2
entrypoints==0.3
ez-setup==0.9
flake8==3.7.7
idna==2.8
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==6.0.0
psycopg2==2.8.2
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.3.1
pytz==2019.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.4
urllib3==1.24.2
whitenoise==4.1.2
wrapt==1.11.1
